requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.8
env==0.1.0
eval_type_backport==0.2.0
fastapi==0.115.2
filelock==3.16.1
firebase-admin==6.5.0
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pipenv==2024.0.3
platformdirs==4.3.6
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
rsa==4.9
setuptools==75.1.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.40.0
tabulate==0.9.0
together==1.3.3
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.26.5
Werkzeug==3.0.4
yarl==1.15.5
